requirements.txt
google-cloud-aiplatform[agent_engines, adk]==1.126.0
google-adk
cloudpickle
pydantic
pyyaml
ipykernel
python-dotenv
google-cloud-discoveryengine
urllib3
requests